techjays_intern_expensetracker/requirements.txt
asgiref==3.7.2
Django==4.1.12
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
mongoengine==0.27.0
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
